what I've got so far, I'll be working on Board
Board, HumanPlayer, and GameController pretty self explanatory.  All the
menus will contain JButtons and the such, and extend Menu.  Tile,
RoadNode, and TownNode (I'm calling them towns, not settlements) are the
parts of a board.  RoadNode and TownNode extend Node, which I'm debating
keeping as it's been useless so far.  ButtonListener is my current
attempt to get JButtons working, but will probably be changed a lot /
gotten rid of.